매니저 수정페이지 작업

+++ frontend/src/pages/manager/ManagerOffice.jsx
@@ -47,39 +47,38 @@ const ManagerOffice = () => {
     };
   }, []);
 
-  // 선택된 페이지 데이터
   const handlePageClick = (data) => {
     const selectedPage = data.selected + 1;
     fetchOffices(selectedPage);
   };
 
-  // 탭 누르면 상태 업데이트 (전체, 서비스 중, 대기 중)
   const handleTabClick = (availabilityStatus) => {
     setAvailability(availabilityStatus);
   };
 
-  // 검색 입력값을 저장
   const handleSearchInputChange = (e) => {
     setSearchInput(e.target.value);
   };
 
-  // 검색어 상태 업데이트하고 데이터요청
   const handleSearch = () => {
     setSearchText(searchInput);
   };
 
-  // 신규 등록 페이지로 이동 버튼
   const handleRegisterClick = () => {
     navigate(`/manager/office/register/${no}`);
   };
 
-  // 수정 페이지로 이동 버튼
   const handleUpdateClick = (officeNo) => {
     navigate(`/manager/office/update/${no}/${officeNo}`);
   };
 
-  // 삭제
   const handleDeleteClick = async (officeNo) => {
+    const confirmed = window.confirm("정말 이 오피스를 삭제하시겠습니까?");
+
+    if (!confirmed) {
+      return;
+    }
+
     try {
       await axios.delete(`/manager/office/delete/${officeNo}`, { withCredentials: true });
       alert("오피스가 성공적으로 삭제되었습니다.");
@@ -142,7 +141,7 @@ const ManagerOffice = () => {
                         {office.availability === 1 ? '승인됨' : '대기 중'}
                       </span>
                     </td>
-                    <td>{new Date(office.regDate).toLocaleDateString()}</td>
+                    <td>{new Date(office.reg_date).toLocaleDateString()}</td>
                     <td>
                       <button onClick={() => handleUpdateClick(office.no)}>수정</button>
                       <button onClick={() => handleDeleteClick(office.no)}>삭제</button>
